# swagger/publicProduct.yaml
openapi: 3.0.0
info:
  title: OMS - Public Product API
  version: 1.0.0
  description: Publicly accessible endpoints for browsing and searching products in the Order Management System

tags:
  - name: PublicProducts
    description: Public product discovery and browsing

paths:
  /api/public/products/search:
    get:
      summary: Search products
      tags: [PublicProducts]
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query text
        - in: query
          name: supplier_id
          schema:
            type: string
            format: uuid
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: min_price
          schema:
            type: number
            format: float
        - in: query
          name: max_price
          schema:
            type: number
            format: float
        - in: query
          name: in_stock
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Products search successful

  /api/public/getAllProducts:
    get:
      summary: Get all products
      tags: [PublicProducts]
      parameters:
        - in: query
          name: supplier_id
          schema:
            type: string
            format: uuid
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Products retrieved successfully

  /api/public/getProduct/{id}:
    get:
      summary: Get product by ID
      tags: [PublicProducts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Product retrieved successfully
        404:
          description: Product not found

  /api/public/categories/{category}/products:
    get:
      summary: Get products by category
      tags: [PublicProducts]
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Category products retrieved successfully
        404:
          description: Category not found

  /api/public/categories:
    get:
      summary: Get available categories
      tags: [PublicProducts]
      responses:
        200:
          description: Categories retrieved successfully

  /api/public/suppliers/{supplierId}/products:
    get:
      summary: Get products by supplier
      tags: [PublicProducts]
      parameters:
        - in: path
          name: supplierId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Supplier products retrieved successfully
        404:
          description: Supplier not found